+++ tests/Rector/Dynamic/AnnotationReplacerRector/Correct/correct.php.inc
