+++ tests/Rector/Dynamic/AnnotationReplacerRector/Wrong/wrong.php.inc
