+++ tests/Rector/Dynamic/AnnotationReplacerRector/config/rector.yml
